+++ tests/Mock/ExclusionStrategies/GsonMockExclusionStrategyMock.php
@@ -7,6 +7,7 @@
 namespace Tebru\Gson\Test\Mock\ExclusionStrategies;
 
 use Tebru\Gson\ClassMetadata;
+use Tebru\Gson\ExclusionData;
 use Tebru\Gson\ExclusionStrategy;
 use Tebru\Gson\PropertyMetadata;
 use Tebru\Gson\Test\Mock\GsonMock;
@@ -34,9 +35,10 @@ class GsonMockExclusionStrategyMock implements ExclusionStrategy
      * Return true if the property should be ignored
      *
      * @param PropertyMetadata $propertyMetadata
+     * @param ExclusionData $exclusionData
      * @return bool
      */
-    public function shouldSkipProperty(PropertyMetadata $propertyMetadata): bool
+    public function shouldSkipProperty(PropertyMetadata $propertyMetadata, ExclusionData $exclusionData): bool
     {
         if (false === $this->skipProperty) {
             return false;